Installed FOSUserBundle for the entity Customer. Modified config, routing and securty for it. Created the entity Client for the OAuth client instead of Customer and used the entity Customer for the partners of BileMo and the entity User for the users of the partners. Updated all entities for it. Created ResourceAccessForbiddenException and ResourceNotFoundException and updated the controllers. Linked the users at the customers.

+++ BileMo/src/AppBundle/Controller/UserController.php
@@ -18,6 +18,9 @@ use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use AppBundle\Exception\ResourceValidationException;
+use AppBundle\Exception\ResourceNotFoundException;
+use AppBundle\Exception\ResourceAccessForbiddenException;
+
 
 class UserController extends FOSRestController
 {
@@ -60,7 +63,10 @@ class UserController extends FOSRestController
      */
     public function listAction(ParamFetcherInterface $paramFetcher)
     {
+        $customer = $this->getUser();
+
         $pager = $this->getDoctrine()->getRepository('AppBundle:User')->search(
+            $customer,
             $paramFetcher->get('keyword'),
             $paramFetcher->get('order'),
             $paramFetcher->get('limit'),
@@ -74,6 +80,9 @@ class UserController extends FOSRestController
      * @param User $user
      * @return User
      *
+     * @throws ResourceNotFoundException
+     * @throws ResourceAccessForbiddenException
+     *
      * @Rest\Get(
      *     path = "/api/users/{id}",
      *     name = "app_user_show",
@@ -84,8 +93,18 @@ class UserController extends FOSRestController
      *     serializerGroups = {"detail_user"}
      * )
      */
-    public function showAction(User $user)
+    public function showAction(User $user=null)
     {
+        $customer = $this->getUser();
+
+        if (empty($user)){
+            throw new ResourceNotFoundException('This resource doesn\'t exist.');
+        }
+
+        if($customer !== $user->getCustomer()){
+            throw new ResourceAccessForbiddenException('You don\'t have the permission to access to this resource.');
+        }
+
         return $user;
     }
 
@@ -122,8 +141,13 @@ class UserController extends FOSRestController
             throw new ResourceValidationException($message);
         }
 
-        $user->setDateCreation(new \DateTime());
+        $customer = $this->getUser();
 
+        $user
+            ->setDateCreation(new \DateTime())
+            ->setIsActive(true)
+            ->setCustomer($customer)
+        ;
         $passwordEncoder = $this->get('security.password_encoder');
 
         $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
@@ -131,20 +155,42 @@ class UserController extends FOSRestController
 
         $em = $this->getDoctrine()->getManager();
 
+        //$listAddresses = $user->getAddresses();
+        //foreach ($listAddresses as $address)
+        //{
+        //    $user->addAddress($address);
+        //    $address
+        //        ->setDateCreation(new \DateTime())
+        //        ->setIsActive(true)
+        //    ;
+        //}
+
         foreach($user->getAddresses() as $address) {
-            $em->persist($address);
+            $address->setUser($user);
+            $address
+                ->setDateCreation(new \DateTime())
+                ->setIsActive(true)
+            ;
+        //    $em->persist($address);
         }
 
-        var_dump($user);die();
+        //var_dump($user);die();
         $em->persist($user);
         $em->flush();
 
-        return $this->view($user, Response::HTTP_CREATED, ['Location' => $this->generateUrl('app_user_show', ['id' => $user->getId(), UrlGeneratorInterface::ABSOLUTE_URL])]);
+        return $this->view(
+            $user,
+            Response::HTTP_CREATED,
+            ['Location' => $this->generateUrl('app_user_show', ['id' => $user->getId(), UrlGeneratorInterface::ABSOLUTE_URL])]
+        );
     }
 
     /**
      * @param User $user
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @throws ResourceNotFoundException
+     * @throws ResourceAccessForbiddenException
      *
      * @Rest\Delete(
      *     path="/api/users/{id}",
@@ -153,10 +199,20 @@ class UserController extends FOSRestController
      * )
      * @Rest\View(StatusCode=200)
      */
-    public function deleteAction(User $user)
+    public function deleteAction(User $user=null)
     {
+        if (empty($user)){
+            throw new ResourceNotFoundException('This resource doesn\'t exist.');
+        }
+
         $em = $this->getDoctrine()->getManager();
         $userdelete = $em->getRepository('AppBundle:User')->find($user);
+
+        $customer = $this->getUser();
+
+        if($customer !== $user->getCustomer()){
+            throw new ResourceAccessForbiddenException('You don\'t have the permission to access to this resource.');
+        }
 
         $em->remove($userdelete);
         $em->flush();
